Add tuple conversions, cut as_tuple and as_cons_or_error
This change complements 6f4ac71065. I wanted to make it possible to
destructure arbitrary conses, as in

  let (a, (b, c)) = obj.into();

but I don't think that's possible. A little destructuring is possible
though, and this replaces `as_tuple` with it. `as_cons_or_error` is
also cut along the way, in favor of `From`/`into`.

A few style changes have also been made wrt destructuring, like

  -        self.valcell.as_cons_or_error().cdr()
  +        let (_, d) = self.valcell.into();
  +        d

I think this is a little more in keeping with the Rust mindset of
exhaustive matching -- the caller explicitly acknowledges that there
is another value in the cons cell, and that that value is being thrown
away.